requirements.txt
numpy
torch
tensorboard
python-socketio
h5py